Improving the download speed
