+++ package.json
@@ -15,6 +15,7 @@
     "gulp-rename": "^2.0.0",
     "gulp-sourcemaps": "2.6.5",
     "gulp-svgstore": "^7.0.1",
+    "gulp-uglify": "^3.0.2",
     "gulp-webp": "^4.0.1",
     "stylelint": "13.6.1",
     "stylelint-config-htmlacademy": "0.1.4"
@@ -45,6 +46,7 @@
   "dependencies": {
     "csso": "^4.0.3",
     "del": "^5.1.0",
-    "grunt-cli": "^1.3.2"
+    "grunt-cli": "^1.3.2",
+    "gulp-htmlmin": "^5.0.1"
   }
 }
