+++ source/catalog.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <title>HTML Academy: Мишка</title>
-    <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="css/styles.min.css">
   </head>
   <body class="page__body">
 
@@ -193,7 +193,7 @@
             По просьбам наших любимых фолловеров
             мы сняли для вас подробное видео о том, как появляются наши товары.
           </p>
-          <a class="production__button button" href="#">Сделать заказ</a>
+          <a class="production__button button product-week__button-order">Сделать заказ</a>
         </div>
       </section>
     </main>
@@ -240,6 +240,26 @@
         </div>
       </div>
     </footer>
-    <script src="js/menu.js"></script>
+    <section class="modal-size">
+      <p class="modal-size__title">
+        Добавить в корзину
+      </p>
+      <p class="modal-size__text">
+        Выберите размер:
+      </p>
+      <ul class="modal-size__buttons-list">
+        <li>
+          <button class="modal-size__button" type="button">S</button>
+        </li>
+        <li>
+          <button class="modal-size__button" type="button">M</button>
+        </li>
+        <li>
+          <button class="modal-size__button" type="button">L</button>
+        </li>
+      </ul>
+      <button class="modal-size__button-sub" type="submit">Добавить</button>
+    </section>
+    <script src="js/scripts.min.js"></script>
   </body>
 </html>
