+++ source/form.html
@@ -4,7 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <title>HTML Academy: Мишка</title>
-    <link rel="stylesheet" href="css/style.css">
+    <link rel="stylesheet" href="css/styles.min.css">
   </head>
   <body class="page__body">
 
@@ -257,6 +257,6 @@
         </div>
       </div>
     </footer>
-    <script src="js/menu.js"></script>
+    <script src="js/scripts.min.js"></script>
   </body>
 </html>
