+++ source/index.html
@@ -4,7 +4,8 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <title>HTML Academy: Мишка</title>
-    <link rel="stylesheet" href="css/style.css">
+    <link rel="preload" href="css/styles.min.css" as="style">
+    <link rel="stylesheet" href="css/styles.min.css">
   </head>
   <body class="page__body">
 
@@ -157,7 +158,7 @@
           <img class="product-week__img" src="img/triple-set-mobile@1x.jpg" srcset="img/triple-set-mobile@2x.jpg 2x" width="260" height="315" alt="Набор из трёх корзинок">
         </picture>
         <p class="product-week__price">Цена: 990 руб.</p>
-        <a class="product-week__button-order button" href="#">Заказать</a>
+        <a class="product-week__button-order button">Заказать</a>
       </section>
       <section class="advantages">
         <h2 class="advantages__title">Коротко о нас:</h2>
@@ -287,6 +288,6 @@
       </ul>
       <button class="modal-size__button-sub" type="submit">Добавить</button>
     </section>
-    <script src="js/menu.js"></script>
+    <script src="js/scripts.min.js"></script>
   </body>
 </html>
