--- source/js/README
@@ -1,9 +1,0 @@
-Папка для скриптов.
-
---
-
-Структура файлов:
-
-js/
-| - app.js
-| - [другие *.js файлы]
